+++ examples/sysctl_memory.rs
@@ -3,11 +3,10 @@
 //! See https://man.openbsd.org/sysctl.2
 //! and https://docs.rs/libc/latest/libc/fn.sysctl.html
 
-use i3status_again::sensors::sysctl::openbsd::{CTL_VM, VM_UVMEXP, sysctl_uvmexp, uvmexp};
+use i3status_again::sensors::sysctl::openbsd::{sysctl_uvmexp, uvmexp};
 
 fn main() {
-    let mib = [CTL_VM, VM_UVMEXP];
-    let mem: uvmexp = sysctl_uvmexp(&mib).unwrap();
+    let mem: uvmexp = sysctl_uvmexp().unwrap();
 
     println!("Page size: {}", mem.pagesize);
     println!("Number of pages: {}", mem.npages);