+++ src/blocks/mod.rs
@@ -11,11 +11,13 @@ pub mod battery;
 pub mod cpu_freq;
 pub mod cpu_temp;
 pub mod datetime;
+pub mod memory;
 
 pub use crate::blocks::battery::BatteryBlock;
 pub use crate::blocks::cpu_freq::CpuFreqBlock;
 pub use crate::blocks::cpu_temp::CpuTempBlock;
 pub use crate::blocks::datetime::DateTimeBlock;
+pub use crate::blocks::memory::MemoryBlock;
 
 use crate::bar::BlockOutput;
 
@@ -28,6 +30,7 @@ pub enum BlockType {
     CpuFreq(CpuFreqBlock),
     CpuTemp(CpuTempBlock),
     DateTime(DateTimeBlock),
+    Memory(MemoryBlock),
 }
 
 impl BlockType {
@@ -38,6 +41,7 @@ impl BlockType {
             Self::CpuFreq(c) => c.get_output(),
             Self::CpuTemp(c) => c.get_output(),
             Self::DateTime(d) => d.get_output(),
+            Self::Memory(m) => m.get_output(),
         }
     }
 }