+++ CHANGELOG.md
@@ -9,12 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- A block `memory`for OpenBSD, that displays the amount of percentage of used
+  memory.
 - An `interval` option in the configuration to let the user decide how often the
   bar updates its information. Default to 1 second.
-- A block "cpu_freq" for OpenBSD, that displays the frequency of the CPU in MHz
+- A block `cpu_freq` for OpenBSD, that displays the frequency of the CPU in MHz
   or GHz. If several CPUs are combined in one block, it's possible to display the
   average or maximum frequency.
-- A block "cpu_temp" for OpenBSD, that displays the temperature of the CPU in
+- A block `cpu_temp` for OpenBSD, that displays the temperature of the CPU in
   Celsius or Fahrenheit.
 - Documentation in the [./docs](./docs) directory to explain the choices that
   have been made, the project architecture, and describe how the blocks work.